Renamed the language package to schema
