+++ databases/graphs/applications/sdk.go
@@ -1,8 +1,8 @@
 package applications
 
 import (
-	"github.com/steve-care-software/steve/databases/graphs/domain/languages/connections/links"
 	"github.com/steve-care-software/steve/databases/graphs/domain/routes"
+	"github.com/steve-care-software/steve/databases/graphs/domain/schemas/connections/links"
 )
 
 // Application represents the graphdb application
