+++ databases/graphs/domain/routes/sdk.go
@@ -1,6 +1,6 @@
 package routes
 
-import "github.com/steve-care-software/steve/databases/graphs/domain/languages/connections/links/references"
+import "github.com/steve-care-software/steve/databases/graphs/domain/schemas/connections/links/references"
 
 // Routes represents routes
 type Routes interface {
