+++ databases/graphs/domain/schemas/connections/headers/names/cardinalities/cardinality.go
