+++ databases/graphs/domain/schemas/connections/headers/names/sdk.go
@@ -1,6 +1,6 @@
 package names
 
-import "github.com/steve-care-software/steve/databases/graphs/domain/languages/connections/headers/names/cardinalities"
+import "github.com/steve-care-software/steve/databases/graphs/domain/schemas/connections/headers/names/cardinalities"
 
 // NewBuilder creates a new builder
 func NewBuilder() Builder {
