+++ databases/graphs/domain/schemas/connections/links/references/builder.go
