+++ databases/graphs/domain/schemas/connections/links/references/externals/builder.go
