+++ databases/graphs/domain/schemas/connections/suites/expectations/builder.go
@@ -3,7 +3,7 @@ package expectations
 import (
 	"errors"
 
-	"github.com/steve-care-software/steve/databases/graphs/domain/languages/connections/links/references"
+	"github.com/steve-care-software/steve/databases/graphs/domain/schemas/connections/links/references"
 )
 
 type builder struct {
