+++ databases/graphs/domain/schemas/connections/suites/sdk.go
@@ -1,8 +1,8 @@
 package suites
 
 import (
-	"github.com/steve-care-software/steve/databases/graphs/domain/languages/connections/links/references"
-	"github.com/steve-care-software/steve/databases/graphs/domain/languages/connections/suites/expectations"
+	"github.com/steve-care-software/steve/databases/graphs/domain/schemas/connections/links/references"
+	"github.com/steve-care-software/steve/databases/graphs/domain/schemas/connections/suites/expectations"
 )
 
 // NewBuilder creates a new builder instance
