+++ databases/graphs/domain/schemas/helper.go
